reorganize yandex hooks code to avoid docs issue
