+++ tests/test_project_structure.py
@@ -47,7 +47,8 @@ MISSING_TEST_FILES = {
     'tests/providers/oracle/operators/test_oracle.py',
     'tests/providers/qubole/hooks/test_qubole.py',
     'tests/providers/samba/hooks/test_samba.py',
-    'tests/providers/sqlite/operators/test_sqlite.py'
+    'tests/providers/sqlite/operators/test_sqlite.py',
+    'tests/providers/yandex/hooks/yandex.py'
 }
